requirements.txt
playwright>=1.47.0
PyYAML>=6.0
httpx>=0.27.0
